Add face (#9)
This commit adds the splunk_hec puppet face/app allowing for a cat json | puppet splunk_hec like workflow. The first functionality of this code is to enable sending pe metrics data to Splunk using the current CS best practices for collecting the CS data.

This will need changes in the puppet metrics collector module managed by CS to enable that specific workflow: puppetlabs/puppetlabs-puppet_metrics_collector#6

This also adds the transaction_uuid to the fact event submission, so correlation of facts created, the catalog, and the report are now possible.

This introduces an epochtime function to ensure that across all of our event submission code, we're always generating a timestamp of the same precision and in the same way.