Adds ability to manage route_file setting
By setting manage_routes = true in the module the user can now set the
routes.yaml to use the splunk_hec_routes.yaml file generated by this
module. The default parameter values enable the splunk_hec as the cache
terminus for facts, since it can no longer be used as the primary terminus
and expect for Puppet Enterprise to continue to operate properly.