PIE-178 Multiple Metrics in stdin
Before this we assumed we would receive parsable json on standard in from the metrics executable regardless of number of systems involved.

Instead it would output a json hash for each server and not attempt to combine them into a valid array. This converts any hash received on STDIN into an array of hashes (even if its just an array of 1).