Enable Customized Reporting
This update includes the ability to customize how much data is sent in each puppet report. Before this update one required to have the Splunk App working in order to see agent logs or resource events.

This update also revises the documentation for readability and to include the new settings.

Changes to metadata.json to indicate planned 0.8.0 release.