Enable extra data (#38)
* PIE-178 Multiple Metrics in stdin

Before this we assumed we would receive parsable json on standard in from the metrics executable regardless of number of systems involved.

Instead it would output a json hash for each server and not attempt to combine them into a valid array. This converts any hash received on STDIN into an array of hashes (even if its just an array of 1).

* Adds a task that allows a user to revoke Tokens

This is a fix for behavior introduced by the report viewer on the splunk side that could leave stale tokens around for a long time.

* Removes the possibility of user input

Since the Splunk App labeled the tokens, this only revokes the tokens created by that.

No longer needed using input, this becomes a much safer script to run (and wont kick a user out of a console session if they used their own account to generate the token)

* Reverts to username

Tokens can only be revoked by label by user who created them, since we are using the api_user in this instance, we should just use the username

String is handled better and embedded in json post instead of just inline command

* Adds configuration options for loading extra data to reports

* Fix logic / names for settings

* fix logic

* load settings properly

* load settings properly

* just send resources with events

* remove redundant call for resource events

* remove redundant call for resource events

* add logs for corrective change

* fix nil issue for catalog_uuid

* Enable Customized Reporting

This update includes the ability to customize how much data is sent in each puppet report. Before this update one required to have the Splunk App working in order to see agent logs or resource events.

This update also revises the documentation for readability and to include the new settings.

Changes to metadata.json to indicate planned 0.8.0 release.

* Adds note on Puppet Metrics Collector

* remove accidental commit

* fix lint errors