(PIE-296) Add option to dynamically calculate the 'reports' setting
This option is enabled when $reports == '' for backwards compatibility.
The option's useful because it no longer requires the user to specify
_all_ of their report processors in the $reports param. This leads to a
better UX especially when the user wants to use another report processor
module because they do not have to update the $reports parameter to
include that module's report processor.

Note that the $reports parameter will eventually be deprecated.

Signed-off-by: Enis Inan <enis.inan@puppet.com>