(PIE-281) Revive Splunk Acceptance Tests
This PR migrates much of the acceptance testing infrastructure from the
reporting module such that now a puppet master node is created and
installed on vmpooler with the splunk container running within it.

No new acceptance tests have been added. but the existing tests run once
again.

Additionally the splunk container is started and ready to accept HEC
requests for further acceptance testing. The Puppet Report Viewer
Splunkbase app is installed into the Splunk container so that we can now
send HEC requests to the container and verify that they were indexed
correctly.

A jenkins project has been created.
https://cinext-jenkinsmaster-pipeline-prod-1.delivery.puppetlabs.net/view/PIE%20Team/job/pipeline_puppetlabs-splunk_hec/