(PIE-347) Add task to send arbitrary data to Splunk
This is a very basic task to send data to Splunk other than report data.
A customer has requested the ability to capture an audit log of all Bolt
actions in Splunk. While this is not currently supported, this task will
allow them to send arbitrary data to Splunk and design their own audit
log system by including this task in other plans.