(PIE-328) Standardize Release Actions
Add the release_prep step from the reporting integration which creates
the release PR and updates the release action to be consistent. At some
point we may want to update the name of the password secret from token
to password in github since it actually refers to the puppetlabs account
password. Tokens are not supported with the Puppetlabs user.