(PIE-431) Fix Open Source Compatibility
Prior to this change the values for the resource parameters were
hard coded to match Puppet Enterprise.

This change enables the module to determine if it is in a PE environment
or Open Source, and select the appropriate values to avoid errors.

It also fixes a bug in the `requires` statements to account for path
differences between PE and Open Source.