(MAINT) Expect report files on puppet server to have a .yaml extension.
A test in the acceptance suite expects
/opt/puppetlabs/puppet/cache/reports/ on the puppet server to contain a
report file with a .json extension. This test is now failing because the
report file has a .yaml extension. We don't currently understand what
caused this change, but the reports are being generated without issue
which was the intent of the test. It's not clear that a change to this
module was responsible. This change modifies the expectation of the test
from a .json to a .yaml file.