(MAINT) Fix gem user install bug
The puppetlabs/action-litmus_spec action installs the `bundler` gem
using the `--user-install` flag to the `gem install` command. Doing this
installs the `bundler` gem to a user folder that is not on the PATH.

For some reason this has starting causing CI to fail.

This change switches to a user fork of the action until a fix can be
implemented in the upstream repository.