(MAINT) - Create a testing account to pull apps from Splunkbase
The Splunk container requires Splunkbase credentials to pull the Report
Viewer App from Splunkbase and install it into the container for
testing. This change adds credentials that are only used for this and
can be easily recreated. Any Splunkbase account will do. Password has
been changed on the other account that was here previously which got
some editing permission after the creds were committed here.