(PIE-459) Fix Ruby Loadpath for collector script
When the splunk_hec_collect_api_events.rb script is executed, the
`$LOAD_PATH` variable is not set to allow `require` to find the
`common_events_library` gem even though the module has been installed
into Puppet.

This change asks Puppet where the modules are installed, and finds the
correct module and adds the correct directory to `$LOAD_PATH` so that
requires statement will work.