(MAINT) Default include_api_collection parameter to false
When installing the module users who do not wish to use api collection
will be greeted with a validation error if the include_api_collection
parameter is set to true, but the pe_console, pe_username, and
pe_password parameters are not set. This changes the default to be false
so that exising users who are not using the api collection features will
not run into this validation.