(MAINT) Adjust code to work with Common Events
Common events code has changed slightly, so the names of the methods
and properties that need to be called need to be fixed in here to
account for it.

There is also much less JSON parsing to do here in this file because it
is now handled in the library code.