(MAINT) Improve acceptance tests
Instead of using a time based splunk query to see how many events the
script sent to splunk, we switch to a `last n` events filter and just
test to see if we got the events we expected to get.

Switch from executing `master.run_cmd` five times, call it a single time
and execute a command that does the looping on the remote side.

Remove the need for instance variables and `before(:all)` blocks that
Rubocop doesn't like.