(MAINT) Make Splunk Releasable
A lot of proof of concept work for sending Activities API events from
PE to Splunk was done on the main branch of this repo. There has been
a major change in strategy such that this work was no longer valid.

Unfortunately there we some contributions and bug fixes to this module
that got interwoven with the Events sending code. We can't release the
module with all of the dead event sending code still in it, but that
also meant that we could not release the legitimate bug fixes and
improvements.

This PR removes all of the Events sending code while retaining the
interwoven bug fixes and improvements, without re-writing the history
of the main branch so that the module can be brought to a releasable
state.