(PIE-606) Pin Litmus and Provision versions
This workaround pins the versions of the provision module and
puppet_litmus gem to work with the reporting_integration. Specifically
it restores functionality to provision with vmpooler and copy the
inventory file.

Furthermore, we remove several references to the common events gem
because it will be removed shortly and passing tests is a higher
priority.

Provision module is set to a specific SHA.
Puppet_litmus is set to version 0.26.0 or earlier (lowerbound of 0.18).