(PIE-818) Allow Ignore System Certificate Store
This change allows a user to provide a ca certificate file that
overrides the system trusted certificate store entirely.

Prior to this change, a ca certificate file provided to the `ssl_ca`
parameter was only a supplement to the system certificates store. This
could cause an issue where the system certificate store contained
invalid certificates and would cause certificate validation of the
splunk_hec end point to fail before the provided `ssl_ca` file was
checked.

This change fixes that bug by ensuring that only the provided file is
used for verification.