(PIE-929) Fix module to work on FIPS servers
Prior to this change, this module could not function on FIPS enabled
host operating systems and FIPS enabled Puppet servers. The reason is
this module consumes the `net/http` and `net/https` gems directly. These
gems cannot function in FIPS environments.

This change allows the module to detect when it's running in a FIPS
environment and switch to using Puppet servers build in HTTP object.

Puppet server consumes a java library called Bouncy Castle that is FIPS
compliant. The Puppet server code base uses Clojure code to create an
object that Ruby code running in JRuby can consume. It keeps a pool of
these objects and this module can now grab an instance of one of these
clients to do HTTP calls.

The drawback to this approach for now is that we lose some custom
functionality when using Puppet server's client. We can no longer
support a standalone ca file for SSL authentication, and we cannot set
user customizable timeouts.

In order to add FIPS functionality in backwards compatible way, we
detect when we're in FIPS mode and use the Puppet server client, but the
rest of the time, for all of the existing users, we continue to use the
`net/http` and `net/https` gems directly.