(PIE-959) Agent Only Node Fact Broken
The implementation of the fact `splunk_hec_agent_only_node` was flawed.
It attempted to see if there was a difference between the `server` fact
and the current nodes name. It turns out this is not a good method
because in a multi master environment, a node might have a different
name from the `server` fact, but it is still a server node that might
run report processors.

This implementation looks at the services configured on the machine to
determine if a puppetserver service has been registered.