PIE-1047 Account for negative indicies
pe_event_forwarding will soon have the ability to disable rbac event
collection for performance reasons. The is denoted in the index file
with a -1 value. We need to cover the case where rbac events are
enabled in splunk_hec but disabled in pe_event_forwarding. This check
will skip indicies with negative values.